refactor: 스타일, 타입 파일 분리
